Trying more for the SR 595
Trying to use the ShiftRegister74HC595.h library to get the shift register working with the ToFs.
I don't believe that it's possible to get it working with the ToFs, as the SR is only able to do HIGH or LOW on each pin that it has, but doesn't have an address for those pins that is assignable to the XSHUT pins for the ToFs. The SR can be used for things like an LED or a sound trigger or motor, where it just needs to be HIGH or LOW to activate, and doesn't require the pins to be explicitly assigned.