Adding LED strip and comments.
Big changes!

Added LED strip and split the array into sections to be filled in each time a ToF detects an object within a certain range.

Organized ToF loop, added functions for efficiency and organization.

Added comments for clarity.

Now works as when the first ToF detects an object within a certain range, it fills in the corresponding section of the LED strip, then moves to monitoring the next ToF, waiting to detect an object within that range so it can turn on the corresponding section of the LED strip. When it has gone through the TOTAL_TOFS, then the game is done.