Adjusting the distance
Adjusting the distance to detect an object within to see if it helps to see through the clear plastic tube. It can't see through the clear plastic tube because of the way the light is bending around the tube, creating too much ambient feedback to it to see through like a normal sheet of clear plastic.